sys/vfs: Drop dependency to posix headers
Our POSIX compat layer provides POSIX compatible VFS stats based on
RIOT VFS module. Consequently, the POSIX layer should depend on VFS,
rather than VFS depending on POSIX.

This moves a struct definition to the VFS module headers and lets the
POSIX compatibility layer include `vfs.h`, rather than `vfs.h` including
the POSIX compat headers.
