+++ sys/include/vfs.h
@@ -64,7 +64,11 @@
 #endif
 #include <sys/stat.h> /* for struct stat */
 #include <sys/types.h> /* for off_t etc. */
-#include <sys/statvfs.h> /* for struct statvfs */
+
+#ifdef CPU_NATIVE
+/* on native, use system's struct statvfs definition */
+#include <sys/statvfs.h>
+#endif
 
 #include "sched.h"
 #include "clist.h"
@@ -338,6 +342,43 @@ extern const vfs_file_ops_t mtd_vfs_ops;
  * @brief   File system always wants the full VFS path
  */
 #define VFS_FS_FLAG_WANT_ABS_PATH   (1 << 0)
+
+#ifndef CPU_NATIVE
+/* on native, use system's struct statvfs definition */
+
+/**
+ * @brief   File system information
+ *
+ * @details This structure is intentionally compatible with POSIX's
+ *          `struct statvfs` to avoid conversion when accessing VFS stats
+ *          through the POSIX compatibility layer
+ */
+struct statvfs {
+    unsigned long f_bsize;   /**< File system block size. */
+    unsigned long f_frsize;  /**< Fundamental file system block size. */
+    fsblkcnt_t f_blocks;     /**< Total number of blocks on file system in
+                                  units of @c f_frsize. */
+    fsblkcnt_t f_bfree;      /**< Total number of free blocks. */
+    fsblkcnt_t f_bavail;     /**< Number of free blocks available to
+                                  non-privileged process. */
+    fsfilcnt_t f_files;      /**< Total number of file serial numbers. */
+    fsfilcnt_t f_ffree;      /**< Total number of free file serial numbers. */
+    fsfilcnt_t f_favail;     /**< Number of file serial numbers available to
+                                  non-privileged process. */
+
+    unsigned long f_fsid;    /**< File system ID. */
+    unsigned long f_flag;    /**< Bit mask of f_flag values. */
+    unsigned long f_namemax; /**< Maximum filename length. */
+};
+
+/**
+ * @brief   Enumeration of values for use in struct statvfs::f_flag
+ */
+enum {
+    ST_RDONLY = 1,          /**< Mount read-only. */
+    ST_NOSUID = 2,          /**< Ignore suid and sgid bits. */
+};
+#endif /* CPU_NATIVE */
 
 /**
  * @brief A file system driver
